backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
python-multipart==0.0.6
qdrant-client>=1.7.1
openai>=1.3.5
python-dotenv==1.0.0
pypdf==3.17.1
python-docx==1.1.0
requests==2.31.0
pydantic>=2.0
torch
torchvision
transformers>=4.45.0
qwen-vl-utils
Pillow
accelerate
numpy
